Move feign dtos to common lib
