+++ cab-app-rides-service/src/main/java/by/arvisit/cabapp/ridesservice/client/DirectDriverClient.java
@@ -4,7 +4,7 @@ import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 
-import by.arvisit.cabapp.ridesservice.dto.DriverResponseDto;
+import by.arvisit.cabapp.common.dto.driver.DriverResponseDto;
 
 @FeignClient(name = "cab-app-driver-service", url = "${spring.settings.cab-app-driver-service.uri}",
         configuration = CabAppFeignClientConfiguration.class)
