+++ cab-app-rides-service/src/main/java/by/arvisit/cabapp/ridesservice/client/DriverClient.java
@@ -1,6 +1,6 @@
 package by.arvisit.cabapp.ridesservice.client;
 
-import by.arvisit.cabapp.ridesservice.dto.DriverResponseDto;
+import by.arvisit.cabapp.common.dto.driver.DriverResponseDto;
 
 public interface DriverClient {
