+++ cab-app-rides-service/src/main/java/by/arvisit/cabapp/ridesservice/client/PassengerClient.java
@@ -1,6 +1,6 @@
 package by.arvisit.cabapp.ridesservice.client;
 
-import by.arvisit.cabapp.ridesservice.dto.PassengerResponseDto;
+import by.arvisit.cabapp.common.dto.passenger.PassengerResponseDto;
 
 public interface PassengerClient {
